+++ ACFG_gestionChataigneraie/src/Controller/personneController.php
@@ -47,11 +47,13 @@ class personneController extends AbstractController {
     /**
      * @Route("supprimerPersonne/{id}", name="supprimerPersonne")
      */
-    function supprimerPersonne(ManagerRegistry $doctrine, $id) {
-        $entityManager = $doctrine->getManager();
-        $personne = $entityManager->getRepository(Personne::class)->find($id);
-        $entityManager->remove($personne);
-        $entityManager->flush($personne);
+    function supprimerPersonne(ManagerRegistry $doctrine, $id, Request $request) {
+        if ($request->getSession()->get('admin')) {
+            $entityManager = $doctrine->getManager();
+            $personne = $entityManager->getRepository(Personne::class)->find($id);
+            $entityManager->remove($personne);
+            $entityManager->flush($personne);
+        }
         
         return $this->redirect($this->generateUrl('listePersonnes'));
     }
